+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-04-24 20:10:33
-// Commit: fix(i18n): prevent double tap duplicate action in i18n
+// Updated at: 2026-04-25 09:14:57
+// Commit: test(i18n): implement widget integration tests for i18n
 
 class i18nComponentState {
-  final String commitRef = '645';
-  final String lastMessage = 'fix(i18n): prevent double tap duplicate action in i18n';
-  final DateTime timestamp = DateTime.parse('2026-04-24T20:10:33+05:00');
+  final String commitRef = '654';
+  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
+  final DateTime timestamp = DateTime.parse('2026-04-25T09:14:57+05:00');
 }